Kick Swipe and Enemy Animation Adjustments
I added a KICK SWIPE to the player aswell as attempt to fix some issues with the Shieldud animations. Unity is a bit limiting in what I could do but I managed to remove some of the slides in the attack.

I also made an Shieldud prefab with said alterations