Enemy Animations
Sorry for the late update but here's all the enemy animations for Shieldud.
I also included some like walking to allow for the possibility for patrols in certain areas.
Please test and tell me if, when the model is applied in the world if the animations are at the right height. As for some reason Unity makes the initial height of an animation constant trough the rest of it.