Shieldud Animations UPDATED
I switched the prefab for Shieldud alongside 3 extra animations. 2 for a SLAP combo (both can be done in succession OR one can be played alone as they return to idle)
And another for Falling as we may end up having holes in our map that could result in the enemy falling to it's death.