Map Prototype Version B
Just a push to the my Scene Version of the Map. I'm a bit worried about the quality of the mesh but I'm still a bit far from the final product.

It's especially chaotic due to it not having the  mirrored 3 towers and the actual prison really in mind but now it should be going in the right direction? Ehh