Some fixes...There's ALOT to do .-.
I attempted to fix the materials on the characters. Manage to make a working textured prefab of the Player and Shieldud.
Scyrent will need to be reuploaded as it's fbx has too many obstructions.
I attempted to mess with the map as well, added some things made some lit and unlit prefabs but... there's a lot of work to do still.